fix: homing bullets now target enemies for C# Player (Issue #781)
Root cause: `_is_player_bullet()` in bullet.gd used case-sensitive string
check `resource_path.contains("player")`. For C# Player, the script path
is `res://Scripts/Characters/Player.cs`, which contains "Player" (capital P)
but NOT "player" (lowercase). This caused `_is_player_bullet()` to always
return false, so `_apply_homing_steering()` never ran for pistol bullets.

Additionally, `set_shooter_id()` was called with `(int)GetInstanceId()` —
a 32-bit truncation of a 64-bit ulong — causing `instance_from_id()` to
return null for high instance IDs, which also broke the player check.

Fixes:
- `_is_player_bullet()`: use `is_in_group("player")` — consistent with the
  rest of the codebase, works for both C# and GDScript Player nodes.
- `set_shooter_id` calls: cast to `(long)` instead of `(int)` to preserve
  the full 64-bit instance ID.

+++ Scripts/AbstractClasses/BaseWeapon.cs
@@ -444,7 +444,7 @@ public abstract partial class BaseWeapon : Node2D
             var owner = GetParent();
             if (owner != null)
             {
-                bullet.Call("set_shooter_id", (int)owner.GetInstanceId());
+                bullet.Call("set_shooter_id", (long)owner.GetInstanceId());
             }
             bullet.Call("set_shooter_position", GlobalPosition);
         }
